Topic yoon

+++ src/main/java/www/project/domain/StarVO.java
@@ -9,8 +9,8 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class StarVO {
-    private String email;
-    private int rate;
-    private long mediaId;
-    private String date;
+    private String email; // 유저아이디
+    private float rate; // 별점
+    private long mediaId; // 무비아이디
+    private String date; // 별점 준 시간
 }
